requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
mangum
pillow